Correct gradient log likelihood
Correct mistake in the gradient of the log likelihood, that needs to take into account the gradient of the determinant of the Jacobian. Add a method that compute the gradient of the log of the determinant of the jacobian. Add a few comments.
Add monitoring convergence using mean computation in post_process