Correct Monte Carlo Propagation
We corrected the propagation using the Monte Carlo method. It was not possible to use it since the addition of PCE method by Lacroix. Now, if the "emulator" keyword ( in the input file) in set to "None", then direct MC propagation is used.

We added an example 'pyrolysis_mass_loss_propagation' for propagation using Monte Carlo and using a different model than for the inference.

We corrected the reading of input data files for inference. We don't need to specify the "_0" suffix anymore, we just need the name of the input file without the ".csv" extension when n_runs is not specified. The suffixes "_0", "_1", etc are added automatically if n_runs is specified, and the input files must be named accordingly.

We added the post process for propagation. It reads the output files created by the Monte Carlo propagation. It does not yet consider output from PCE method. For that, we still need to use the "check.py" function, see for instance in "example/pyrolysis".