Generalise IO_util, compute log gradient and evaluation of distribution
Change the input/output utilities. Both path and fileID are in a IO_util variable, such that only this one has to be passed in argument of other classes/functions. This also allowed to add the folder path "/output/model_eval" where we stored all the model evaluation during sampling of Bayesian posterior.

Correct the computation of the gradient of the log-likelihood. It was tailored to the pyrolysis cases, where you can have multiple species. Now it is general. The gradient must be a dict with the only keys the uncertain parameter names.

The direct evaluation of a distribution is now computed with the function "compute_value_no_reparam". This is useful to avoid computing the posterior distribution in the reparameterized space when there is one reparameterization.