Improve covariance update calculation
Improve the covariance update. The condition for the starting iteration, update and end iteration are better defined. Also the inverse of L_c is now computed only at the update of the covariance, and only once (the transpose of the inverse being the inverse of the transpose).

Also improve computation of the log value of the Bayesian posterior. We separate the condition on the value of prior_log_value and log_like_val, such that if the prior is zero (e.g. out of bounds), then the model will not be run.