Implement pathlib for Bayesian inference module
Remove the use of os package. The manipulation of the paths is now performed using pathlib. Remove all the path in "Windows-style", in order to make the code compatible with other operating systems. This is done for Bayesian inference. Still need to be done for propagation.
Redefine the multiple outputs, and clarify them. Remove deprecated outputs.